Remove CoverageReportProcessorBaseCopy.cs
